+++ source/extensions/lightspeed.trex.properties_pane.shared.mod_setup.widget/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.6.0]
+## Removed
+- removed "Open Wizard" button from property panel
+
 ## [1.5.7]
 ## Changed
 - Improved Mod File button labels